requirements.txt
Django==4.2.10
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework==3.14.0
requests==2.31.0
beautifulsoup4==4.12.3
markdown==3.5.1
django-filter==23.5
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
openai==1.12.0
django-allauth==0.57.0
gunicorn==21.2.0
whitenoise==6.5.0
django-storages==1.14.2
boto3==1.34.34
PyGithub==2.1.1
arxiv==2.0.0 